requirements.txt
bunch_py3==2.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
